+++ rstsr-openblas/tests/mod.rs
@@ -1,2 +1,3 @@
 mod test_driver_impl;
+#[cfg(feature = "linalg")]
 mod test_linalg_func;
